add condition for delete content

+++ src/Components/Admin/Item/index.tsx
@@ -10,26 +10,30 @@ import Api from '../../../Services/Api';
 // styles
 import { ItemContainer } from './styles';
 async function handleDeleteContent(id: string) {
-    Api.delete(DELETE, {params: {id}}).then((response) => {
-        if (response.status === 204) {
-            alert('O conteúdo foi deletado com sucesso!')
-            return true
+    const isDelete = window.confirm('Você deseja deletar o conteúdo?')
+
+    if (isDelete) {
+        Api.delete(DELETE, {params: {id}}).then((response) => {
+            if (response.status === 204) {
+                return true
+            }
+        }).catch(err => {
+            if(err.response.status === 404){
+                alert('O conteúdo não existe!')
+            return false
+        }else if(err.response.status === 409) {
+            alert('O conteúdo não foi deletado por um erro de conflito')
         }
-    }).catch(err => {
-        if(err.response.status === 404){
-            alert('O conteúdo não existe!')
-        return false
-    }else if(err.response.status === 409) {
-        alert('O conteúdo não foi deletado por um erro de conflito')
+        
+        else if (err.response.status === 500) {
+            alert('Ocorreu um erro interno, entre em contato com o suporte')
+            return false
+        }else {
+            alert('Erro não informado, entre em contato com o suporte!')
+        }
+        })
     }
     
-    else if (err.response.status === 500) {
-        alert('Ocorreu um erro interno, entre em contato com o suporte')
-        return false
-    }else {
-        alert('Erro não informado, entre em contato com o suporte!')
-    }
-    })
    
 }
